feat: divided router.py to 2 new router
